+++ lib/ca_testing/patches/sample.rb
@@ -2,17 +2,7 @@
 
 module CaTesting
   module Patches
-    class Sample
-      def patch!
-        raise "This is no longer supported" if prevent_usage?
-
-        Kernel.warn("This is now deprecated and should not be used") if deprecate?
-        CaTesting.logger.info("Adding patch: #{self.class}")
-        CaTesting.logger.debug(description)
-        perform
-        CaTesting.logger.info("Patch successfully added.")
-      end
-
+    class Sample < Base
       private
 
       def description
@@ -26,16 +16,8 @@ module CaTesting
         :change_some_behaviour
       end
 
-      def deprecate?
-        Time.new > deprecation_notice_date
-      end
-
       def deprecation_notice_date
         Time.new(2021, 2, 2)
-      end
-
-      def prevent_usage?
-        Time.new > prevent_usage_date
       end
 
       def prevent_usage_date
